backend/requirements.txt
fastapi>=0.127.0,<0.128.0
uvicorn>=0.40.0,<0.41.0
python-dotenv>=1.2.1,<1.3.0
websockets>=15.0.1,<15.1.0
psutil>=7.1.3,<7.2.0
